pyproject.toml
[project]
name = "chess"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "python-chess>=1.2.0",
    "requests>=2.32.5",
    "opencv-python>=4.10.0.0",
    "numpy>=2.0.0",
]
